Inventory code fully implemented! (theoretically)
All the code is now done!... Kind of-
The item code will need to be messed with to implement the usage of healing items etc. and I still need to mess with how bullets will be removed BUT the main part of it is over!

Next hurdle will be to mess with the above-mentioned items that need to be created, but shouldn't be too hard in theory with the set-up code.

Our other problem will likely be with testing the items, since I think player collision is still unfortunately messing with the pick-ups, and I'd also like to be able to map the pick-up system differently for our game! Map it to an interact if possible!

Notable Additions:
- Inventory will persist between scenes!
(I also added this to the player, let me know if there is any duplicate problems or such! Im not sure how you want to deal with the player. Im sure the inventory will be fine tho!)
- I was gonna make a big thing now I forgot im sorry lol its almost 4am

Things to be messed with:
- Actually implementing the items properly
  > attach items to their respective objects? (Might ask Zen for item models maybe, idk how
      necessary that is)
  > Make each item a fitting description and a cool image
- A player health needs to be established for healing items (same for insanity, if we still want a thing for that to be healed etc.)
- Potentially make a key items menu
- Make the UI pretty ehehehe