Added some flare to the ui! Also figured out more problems im gonna mess with
So I added some paper textures to the inventory UI! (though I'll likely tweak them more in the future, especially the background image of the item description area)

In doing that, I found the problem of why my quantities and image sprites aren't popping up! But also not really... Its a spaghetti of a problem!

The quantity text and item image are not getting set to true once there is an item (though the quantity text does change to the right number, it's just not visible to the player), which means something in the code is not enabling that.

But that was already obvious with the items messing up, likely due to naming discrepancies Mel mentioned, and I'm trying to use my brain to fix it (which isn't good cause I'm not that smart with this kind of stuff). But I saw that there is a reason for it within the tutorial I watched, but the reason isn't great, considering it seems to be the root of my problem (potentially).