Item Images Update
Not a whole lot done, I'll do more after a get a few hours sleep... But items go brr! I tried to touch a little bit of the use item again but it absolutely sucks and i dont know where to move it and implement it properly right now sobbing.